+++ src/geneweaver/client/__init__.py
@@ -1,3 +1,4 @@
 """The root of the Geneweaver client module."""
 # ruff: noqa: F401
-from geneweaver.client import auth, config, core, exceptions
+from geneweaver.client import auth, core, exceptions
+from geneweaver.client.core import config
