--- src/geneweaver/client/api/v3/__init__.py
@@ -1,1 +1,0 @@
-"""Api Client for GeneWeaver v3 API."""
