+++ src/geneweaver/client/core/config_class.py
@@ -12,30 +12,21 @@ class Settings(BaseSettings):
     AUTH_ALGORITHMS = ["RS256"]
     AUTH_SCOPES = ["openid", "profile", "email"]
 
-    API_HOST: str = "https://geneweaver.org"
-    API_V1_PATH: str = "/api/v1"
-    API_V2_PATH: str = "/api/v2"
-    API_V3_PATH: str = "/api/v3"
-    API_PATH: Optional[dict] = None
-
-    def api_v3_path(self) -> str:
-        """Construct the API v3 path."""
-        return self.API_HOST + self.API_V3_PATH
-
-    @validator("API_PATH")
-    def validate_api_path(
-        cls: Type["Settings"], v: Optional[dict], values: dict  # noqa: N805
-    ) -> dict:
-        """Construct the API URL if not explicitly set."""
-        if not v:
-            return {
-                "v1": values["API_HOST"] + values["API_V1_PATH"],
-                "v2": values["API_HOST"] + values["API_V2_PATH"],
-                "v3": values["API_HOST"] + values["API_V3_PATH"],
-            }
-        return v
+    API_HOST: str = "https://geneweaver-prod.jax.org"
+    API_PATH: str = "/api"
 
     API_URL: Optional[str] = None
+
+    @validator("API_URL")
+    def validate_api_url(
+        cls: Type["Settings"], v: Optional[str], values: dict  # noqa: N805
+    ) -> str:
+        """Construct the API URL if not explicitly set."""
+        if not v:
+            return values["API_HOST"] + values["API_PATH"]
+        return v
+
+    API_KEY: Optional[str] = None
 
     class Config:
         """Settings configuration."""
