--- src/geneweaver/client/core/config_class_legacy.py
@@ -1,1 +1,0 @@
-"""A legacy version of the GeneWeaver Client configuration module."""
