+++ tests/unit/test_client_config.py
@@ -1,13 +1,13 @@
-"""Unit tests for the geneweaver.client.config module."""
+"""Unit tests for the geneweaver.client.core.config module."""
 # ruff: noqa: ANN001, ANN201
 # TODO: This instance needs to be provided and configured as a fixture so that
 #  it can override which .env files to use for testing.
-from geneweaver.client.config import ClientSettings, settings
+from geneweaver.client.core.config import Settings, settings
 
 
 def test_client_settings_schema():
-    """Test the ClientSettings class."""
-    schema = ClientSettings.schema()
+    """Test the Settings class."""
+    schema = Settings.schema()
     assert "API_HOST" in schema["properties"]
     assert "API_PATH" in schema["properties"]
     assert "API_URL" in schema["properties"]
@@ -18,7 +18,7 @@ def test_client_settings_schema():
 # TODO: This test needs its settings instance to be provided and configured as
 #  a fixture so that we can override which .env files to use for testing.
 def test_client_settings_default():
-    """Test the ClientSettings class."""
+    """Test the Settings class."""
     assert settings.API_HOST == "https://geneweaver.org"
     assert settings.API_PATH == "/api/v2"
     assert settings.API_URL == "https://geneweaver.org/api/v2"
@@ -27,8 +27,8 @@ def test_client_settings_default():
 
 
 def test_client_settings_kwargs():
-    """Test the instantiating ClientSettings class with kwargs."""
-    these_settings = ClientSettings(
+    """Test the instantiating Settings class with kwargs."""
+    these_settings = Settings(
         API_HOST="test_0",
         API_KEY="test_1",
         API_PATH="test_2",
@@ -41,8 +41,8 @@ def test_client_settings_kwargs():
 
 
 def test_client_settings_kwargs_generate_api_url():
-    """Test the instantiating ClientSettings class with kwargs, leaving out API_URL."""
-    these_settings = ClientSettings(
+    """Test the instantiating Settings class with kwargs, leaving out API_URL."""
+    these_settings = Settings(
         API_HOST="test_0",
         API_KEY="test_1",
         API_PATH="test_2",
